requirements.txt
PyQt5
pyqtgraph
pydicom
nibabel
SimpleITK
scikit-image
Pillow
torch
scipy
